refactor(rolldown_binding): split binding_builtin_plugin
