+++ crates/rolldown_binding/src/options/plugin/config/mod.rs
@@ -1,0 +1,3 @@
+mod binding_vite_resolve_plugin_config;
+
+pub use binding_vite_resolve_plugin_config::BindingViteResolvePluginConfig;
