+++ crates/rolldown_binding/src/options/plugin/mod.rs
@@ -1,14 +1,16 @@
-pub mod binding_plugin_context;
+mod binding_builtin_plugin;
+mod binding_callable_builtin_plugin;
+mod binding_plugin_hook_meta;
 mod binding_plugin_options;
 mod binding_transform_context;
+mod config;
 mod js_plugin;
+
+pub mod binding_plugin_context;
 pub mod types;
 
 pub use binding_plugin_options::*;
 pub use js_plugin::*;
-mod binding_builtin_plugin;
-mod binding_callable_builtin_plugin;
-mod binding_plugin_hook_meta;
 
 #[cfg(not(target_family = "wasm"))]
 mod parallel_js_plugin;
